Separated the data models per version
