+++ DemoServer/DataModel/v10/AuthFieldMapping.cs
@@ -1,4 +1,4 @@
-﻿namespace DemoServer.DataModel;
+﻿namespace DemoServer.DataModel.v10;
 
 /// <summary>
 /// The mapping between an AuthField and the field name in the authentication request.
