+++ DemoServer/DataModel/v10/AuthResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace DemoServer.DataModel;
+﻿namespace DemoServer.DataModel.v10;
 
 /// <summary>
 /// The response to an authentication request.
