+++ DemoServer/DataModel/v10/AuthScheme.cs
@@ -1,4 +1,4 @@
-﻿namespace DemoServer.DataModel;
+﻿namespace DemoServer.DataModel.v10;
 
 /// <summary>
 /// Describes one authentication scheme for this data source.
