+++ DemoServer/DataModel/v10/Context.cs
@@ -1,4 +1,4 @@
-﻿namespace DemoServer.DataModel;
+﻿namespace DemoServer.DataModel.v10;
 
 /// <summary>
 /// Matching context returned by the data source as a result of a retrieval request.
