+++ DemoServer/DataModel/v10/RetrievalRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace DemoServer.DataModel;
+﻿namespace DemoServer.DataModel.v10;
 
 /// <summary>
 /// The retrieval request sent by AI Studio.
