+++ DemoServer/Handler/DataSource.cs
@@ -1,11 +1,11 @@
-﻿using DemoServer.DataModel;
+﻿using v10 = DemoServer.DataModel.v10;
 
 namespace DemoServer.Handler;
 
 public static class DataSource
 {
     private const string TAG = "Data Source";
-    private static readonly DataSourceInfo DATA_SOURCE_INFO = new("DEMO: Wikipedia Links", "Contains some links to Wikipedia articles.");
+    private static readonly v10.DataSourceInfo DATA_SOURCE_INFO = new("DEMO: Wikipedia Links", "Contains some links to Wikipedia articles.");
 
     public static void AddDataSourceHandlers(this IEndpointRouteBuilder app)
     {
@@ -18,5 +18,5 @@ public static class DataSource
             .WithName("GetDataSourceInfo");
     }
 
-    private static DataSourceInfo GetDataSourceInfo() => DATA_SOURCE_INFO;
+    private static v10.DataSourceInfo GetDataSourceInfo() => DATA_SOURCE_INFO;
 }
